# Memoize UnitBase._get_physical_type_id