@maxnoe 2022-07-29
I think a changelog entry is not necessary for each of these micro optimizations, but we should probably include a common what's new? Or how do you want to handle this?